requirements.txt
numpy>=1.25
rdkit>=2022.03
pytorch
